requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Image Processing
Pillow>=9.5.0

# Visualization (for Phase 4)
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bar
tqdm>=4.65.0
